requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.3
django-qr-code==3.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.19.5
Pillow==9.1.1
pycparser==2.21
pytz==2022.1
qrcode==7.3.1
segno==1.5.2
sqlparse==0.4.2
